Fix for playlist without #EXTM3U tag

+++ dune_plugin/starnet_edit_playlists_screen.php
@@ -741,7 +741,7 @@ class Starnet_Edit_Playlists_Screen extends Abstract_Preloaded_Regular_Screen im
             }
 
             $contents = file_get_contents($tmp_file, false, null, 0, 512);
-            if ($contents === false || strpos($contents, TAG_EXTM3U) === false) {
+            if ($contents === false || (strpos($contents, TAG_EXTM3U) === false && strpos($contents, TAG_EXTINF) === false)) {
                 unlink($tmp_file);
                 throw new Exception(TR::load('err_bad_m3u_file') . " '$uri'\n\n$contents");
             }
